+++ timerskiplist/timertest.c
@@ -52,13 +52,13 @@ int main(int argc, char **argv)
     timer_skiplist_remove(&priv, timer);
     timer->fn(timer, &priv, timer->userdata, NULL);
 
-    i = rand()%50;
+    i = (size_t)rand()%50;
 
     timer_skiplist_remove(&priv, &periodics[i]);
     periodics[i].time64 += periodics_ud[i].period;
     timer_skiplist_add(&priv, &periodics[i]);
 
-    i = rand()%50;
+    i = (size_t)rand()%50;
     timer_skiplist_remove(&priv, &periodics[i]);
     timer_skiplist_add(&priv, &periodics[i]);
   }
